refactor: move constant file
