+++ src/apis/auth/constant/auth.contant.ts
@@ -1,0 +1,7 @@
+export const SOCIAL_PROVIDERS = {
+  GOOGLE: 'GOOGLE',
+  KAKAO: 'KAKAO',
+} as const;
+
+export type SocialProvider =
+  (typeof SOCIAL_PROVIDERS)[keyof typeof SOCIAL_PROVIDERS];
