+++ src/apis/price-notification/constant/price-notification.constant.ts
@@ -1,0 +1,7 @@
+export const PRICE_NOTIFICATION_TYPES = {
+  TARGET_PRICE: 'target_price',
+  RAPID_FLUCTUATION: 'rapid_fluctuation',
+  DAILY_REPORT: 'daily_report',
+} as const;
+
+export type PriceNotificationType = keyof typeof PRICE_NOTIFICATION_TYPES;
